Create Stat class file
