+++ rpg-npc-creator/rpg-npc-creator/Npc.cs
@@ -61,32 +61,4 @@ namespace rpg_npc_creator
         Stat Statistic { get; }
         Name NpcName { get; }
     }
-    // Add a stat to the NPC
-    public class Stat
-    {
-        int Value;
-
-        public void Set(int NewValue)
-        {
-            this.Value = NewValue;
-            Console.WriteLine("Stat Value is now: " + Value);
-        }
-
-        // Constructors
-        public Stat()
-        {
-            Value = 1;
-            Console.WriteLine("New Stat created with value: " + Value);
-        }
-        public Stat(int IncomingValue)
-        {
-            Value = IncomingValue;
-            Console.WriteLine("New Stat created with value: " + IncomingValue);
-        }
-    }
-    // Add an interface to a stat
-    public interface IStat
-    {
-        int Value { get; }
-    }
 }
