+++ rpg-npc-creator/rpg-npc-creator/rpg-npc-creator.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Npc.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Stat.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
